"updated app.js for debugging"
I was trying to resolve the Ajax issues by nesting functions and I was able to resolve the section that connects the drop down menu to the text box area right below. The text box area populates information about a restaurant with the drop down cuisine category.
Previously, if I would switch selections in the drop down menu, the text box would crash and no longer update, this is no longer an issue. The next problem was introduced when trying to create a bubble plot that would generate a scatter plot for every restaurant within a cuisine category (x axis the rating, an y axis is the average cost).
3 issues here: 1) it crashes when you a new drop down cuisine option is selection and 2) the text box from before would no longer render new information. 3) the bubble plot does not update when a new selection is made, I used a method that should clear out the html area Plotly.deleteTraces("bubble", 0); that was used in the belly button homework assignment, but it does not work here.

Get a 500 error - it might be possible that it is crashing when there is only 1 restaurant for a given category. would need help sorting out this issue.